Slight changes and new skybox
Added double jumping, attempted to fix groundpound (I don't remember if that was a part of my previous push)

When you hold control in middair you ground pound and when you land you enter a slide, I am trying to make it so you enter the slide with a horizontal velocity equivalent to the vertical velocity you gained from the ground pound.

I will leave it for now.